add webapollo functions to retrieve metadata of apollo user, organism and group add permission restrictions to apollo user manager tool * creater apollo user if the gx_user is global admin or instructor * delete apollo user if the gx_user is global admin, or user's creator * add apollo user to group if the gx_user is global admin or group admin or group creator * remove apollo user from group if the gx_user is global admin or group admin or group creator

+++ tools/apollo/apollo_user_manager.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 from __future__ import print_function
 
+import sys
 import argparse
 import random
 import time
@@ -10,10 +11,16 @@ import codecs
 import csv
 from builtins import range, str
 
-from webapollo import WAAuth, WebApolloInstance
+from webapollo import WAAuth, WebApolloInstance, AssertUser, AssertAdmin
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
+def isGroupAdmin(userName, groupName):
+    groupAdmin = wa.groups.getGroupAdmin(groupName)
+    for admin in groupAdmin:
+        if admin['username'] == userName:
+            return True
+    return False
 
 def pwgen(length):
     chars = list('qwrtpsdfghjklzxcvbnm')
@@ -23,7 +30,7 @@ def cleanInput(dict):
     cleanedDict = {k:v.strip() for k,v in dict.items()}
     return cleanedDict
 
-def createApolloUser(user, out):
+def createApolloUser(user, out, gx_user):
     user = cleanInput(user)
     password = pwgen(12)
     time.sleep(1)
@@ -39,17 +46,20 @@ def createApolloUser(user, out):
                        'Email': user['useremail'], 'New Password': password})
         print("Update user %s" % user['useremail'])
     else:
-        returnData = wa.users.createUser(user['useremail'], user['firstname'], user['lastname'], password, role='user')
+        returnData = wa.users.createUser(user['useremail'], user['firstname'], user['lastname'],
+                                         password, role='user', metadata={'creator': gx_user.userId})
         out.writerow({'Operation':'Create User', 'First Name': user['firstname'], 'Last Name': user['lastname'],
                       'Email': user['useremail'], 'New Password': password})
         print("Create user %s" % user['useremail'])
     print("Return data: " + str(returnData) + "\n")
 
 
-def createApolloUsers(users_list, out):
+def createApolloUsers(users_list, out, gx_user):
+    if gx_user.role != 'INSTRUCTOR' and gx_user.role != 'ADMIN':
+        sys.exit(gx_user.role + " is not authorized to create users")
     for user in users_list:
         if user['batch'] == "false":
-            createApolloUser(user, out)
+            createApolloUser(user, out, gx_user)
         elif user['batch'] == "true":
             users = parseUserInfoFile(user['format'], user['false_path'])
             for u in users:
@@ -59,14 +69,18 @@ def createApolloUsers(users_list, out):
                     logger.error("Cannot find firstname in the text file, make sure you use the correct header, see README file for examples.")
                 if not 'lastname' in u:
                     logger.error("Cannot find lastname in the text file, make sure you use the correct header, see README file for examples.")
-                createApolloUser(u, out)
+                createApolloUser(u, out, gx_user)
 
 
-def deleteApolloUser(user, out):
+def deleteApolloUser(user, out, gx_user):
     user = cleanInput(user)
     apollo_user = wa.users.loadUsers(email=user['useremail'])
     if len(apollo_user) == 1:
         userObj = apollo_user[0]
+        # check if gx_user is admin or creator of the apollo_user
+        creatorData = wa.users.getUserCreator(userObj.username)
+        if gx_user.role != 'ADMIN' and creatorData['creator'] != str(gx_user.userId):
+            sys.exit(gx_user.username + " is not authorized to delete user: " + userObj.username)
         returnData = wa.users.deleteUser(userObj)
         out.writerow({'Operation':'Delete User', 'First Name': userObj.firstName, 'Last Name': userObj.lastName,
                       'Email': userObj.username})
@@ -76,20 +90,22 @@ def deleteApolloUser(user, out):
         logger.error("The user %s doesn't exist", user['useremail'])
 
 
-def deleteApolloUsers(users_list, out):
+def deleteApolloUsers(users_list, out, gx_user):
+    if gx_user.role != 'INSTRUCTOR' and gx_user.role != 'ADMIN':
+        sys.exit(gx_user.role + " is not authorized to delete users")
     for user in users_list:
         if user['batch'] == "false":
-            deleteApolloUser(user, out)
+            deleteApolloUser(user, out, gx_user)
         elif user['batch'] == "true":
             users = parseUserInfoFile(user['format'], user['false_path'])
             for u in users:
                 if not 'useremail' in u:
                     logger.error("Cannot find useremail in the text file, make sure you use the correct header, see README file for examples.")
                     exit(1)
-                deleteApolloUser(u, out)
+                deleteApolloUser(u, out, gx_user)
 
 
-def addApolloUserToGroup(user, out):
+def addApolloUserToGroup(user, out, gx_user):
     user = cleanInput(user)
     apollo_user = wa.users.loadUsers(email=user['useremail'])
     groups = wa.groups.loadGroups()
@@ -105,16 +121,20 @@ def addApolloUserToGroup(user, out):
         exit(1)
     userObj = apollo_user[0]
     groupObj = group[0]
+    creatorData = wa.groups.getGroupCreator(groupObj.name)
+    # proceed if gx_user is global admin, group creator or group admin
+    if gx_user.role != 'ADMIN' and not isGroupAdmin(gx_user.username, groupObj.name) and creatorData['creator'] != str(gx_user.userId):
+        sys.exit(gx_user.username + " is not authorized to add user to group: " + groupObj.name)
     returnData = wa.users.addUserToGroup(groupObj, userObj)
     out.writerow({'Operation':'Add User to Group', 'First Name': userObj.firstName, 'Last Name': userObj.lastName,
                   'Email': userObj.username, 'Add to Group': groupObj.name})
     print("Add user %s to group %s" % (userObj.username, groupObj.name))
     print("Return data: " + str(returnData) + "\n")
 
-def addApolloUsersToGroups(users_list, out):
+def addApolloUsersToGroups(users_list, out, gx_user):
     for user in users_list:
         if user['batch'] == "false":
-            addApolloUserToGroup(user, out)
+            addApolloUserToGroup(user, out, gx_user)
         elif user['batch'] == "true":
             users = parseUserInfoFile(user['format'], user['false_path'])
             for u in users:
@@ -124,10 +144,10 @@ def addApolloUsersToGroups(users_list, out):
                 if not 'group' in u:
                     logger.error("Cannot find group in the text file, make sure you use the correct header, see README file for examples.")
                     exit(1)
-                addApolloUserToGroup(u, out)
+                addApolloUserToGroup(u, out, gx_user)
 
 
-def removeApolloUserFromGroup(user, out):
+def removeApolloUserFromGroup(user, out, gx_user):
     user = cleanInput(user)
     apollo_user = wa.users.loadUsers(email=user['useremail'])
     groups = wa.groups.loadGroups()
@@ -143,16 +163,20 @@ def removeApolloUserFromGroup(user, out):
         exit(1)
     userObj = apollo_user[0]
     groupObj = group[0]
+    creatorData = wa.groups.getGroupCreator(groupObj.name)
+    # proceed if gx_user is global admin, group creator or group admin
+    if gx_user.role != 'ADMIN' and not isGroupAdmin(gx_user.username, groupObj.name) and creatorData['creator'] != str(gx_user.userId):
+        sys.exit(gx_user.username + " is not authorized to remove user from group: " + groupObj.name)
     returnData = wa.users.removeUserFromGroup(groupObj, userObj)
     out.writerow({'Operation':'Remove User from Group', 'First Name': userObj.firstName, 'Last Name': userObj.lastName,
                   'Email': userObj.username, 'Remove from Group': groupObj.name})
     print("Remove user %s from group: %s" % (userObj.username, groupObj.name))
     print("Return data: " + str(returnData) + "\n")
 
-def removeApolloUsersFromGroups(users_list, out):
+def removeApolloUsersFromGroups(users_list, out, gx_user):
     for user in users_list:
         if user['batch'] == "false":
-            removeApolloUserFromGroup(user, out)
+            removeApolloUserFromGroup(user, out, gx_user)
         elif user['batch'] == "true":
             users = parseUserInfoFile(user['format'], user['false_path'])
             for u in users:
@@ -162,7 +186,7 @@ def removeApolloUsersFromGroups(users_list, out):
                 if not 'group' in u:
                     logger.error("Cannot find group in the text file, make sure you use the correct header, see README file for examples.")
                     exit(1)
-                removeApolloUserFromGroup(u, out)
+                removeApolloUserFromGroup(u, out, gx_user)
 
 
 def parseUserInfoFile(file_format, filename):
@@ -210,17 +234,22 @@ if __name__ == '__main__':
     csvWriter = csv.DictWriter(outputFile, fieldnames=fieldnames)
     csvWriter.writeheader()
     operations_dictionary = jsonData.get("operations")
+    email = jsonData.get("email")
     wa = WebApolloInstance(args.apollo, args.username, args.password)
-
+    gx_user = AssertUser(wa.users.loadUsers(email=email))
     for operation, users_list in operations_dictionary.items():
         if operation == "create":
-            createApolloUsers(users_list, csvWriter)
+            # proceed if the gx_user is global admin or instructor
+            createApolloUsers(users_list, csvWriter, gx_user)
         elif operation == "delete":
-            deleteApolloUsers(users_list, csvWriter)
+            # proceed if the gx_user is global admin, or user's creator
+            deleteApolloUsers(users_list, csvWriter, gx_user)
         elif operation == "add":
-            addApolloUsersToGroups(users_list, csvWriter)
+            # proceed if the gx_user is global admin or group admin or group creator
+            addApolloUsersToGroups(users_list, csvWriter, gx_user)
         elif operation == "remove":
-            removeApolloUsersFromGroups(users_list, csvWriter)
+            # proceed if the gx_user is global admin or group admin or group creator
+            removeApolloUsersFromGroups(users_list, csvWriter, gx_user)
 
     outputFile.close()
